+++ packages/core-js/features/array/group-by-map.js
@@ -1,5 +1,0 @@
-require('../../modules/es.map');
-require('../../modules/esnext.array.group-by-map');
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Array', 'groupByMap');