+++ packages/core-js/features/array/group-by-to-map.js
@@ -1,0 +1,5 @@
+require('../../modules/es.map');
+require('../../modules/esnext.array.group-by-to-map');
+var entryUnbind = require('../../internals/entry-unbind');
+
+module.exports = entryUnbind('Array', 'groupByToMap');