+++ packages/core-js/features/array/virtual/group-by-map.js
@@ -1,5 +1,0 @@
-require('../../../modules/es.map');
-require('../../../modules/esnext.array.group-by-map');
-var entryVirtual = require('../../../internals/entry-virtual');
-
-module.exports = entryVirtual('Array').groupByMap;