+++ tests/compat/tests.js
@@ -1261,8 +1261,8 @@ GLOBAL.tests = {
   'esnext.array.group-by': function () {
     return [].groupBy;
   },
-  'esnext.array.group-by-map': function () {
-    return [].groupByMap;
+  'esnext.array.group-by-to-map': function () {
+    return [].groupByToMap;
   },
   'esnext.array.is-template-object': function () {
     return Array.isTemplateObject;