+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Added `Number.prototype.toExponential` method polyfill with many different fixes for the most engines
 - [`Array` grouping proposal](https://github.com/tc39/proposal-array-grouping):
   - Moved to the stage 3
-  - Added `Array.prototype.groupByMap` method
+  - Added `Array.prototype.groupByToMap` method
   - Removed `@@species` support
 - Added [change `Array` by copy stage 2 proposal](https://github.com/tc39/proposal-change-array-by-copy):
   - `Array.prototype.toReversed`